+++ .dockerignore
@@ -1,0 +1,4 @@
+.git
+.gitignore
+*.md
+doc
