+++ Dockerfile
@@ -1,27 +1,53 @@
-FROM python:3.7-alpine3.10
+FROM python:3.7-alpine AS builder
 
-WORKDIR /app
+ENV LANG=C.UTF-8
+ENV PYTHONUNBUFFERED=1
 
-COPY *.py *.txt /app/
+# Turns off writing .pyc files
+ENV PYTHONDONTWRITEBYTECODE=1
+
+# Run from virtualenv
+ENV PATH="/venv/bin:$PATH"
+
+ADD ./requirements.txt /requirements.txt
 
 RUN apk upgrade --no-cache \
 	&& apk add --no-cache --virtual .build-deps build-base \
 	gcc gfortran libpng-dev openblas-dev jpeg-dev \
+    && pip install --upgrade pip
+
+RUN python -m venv /venv
+
+RUN pip install --no-cache-dir -r requirements.txt
+
+
+FROM python:3.7-alpine
+
+# Extra python env
+ENV PYTHONDONTWRITEBYTECODE=1 \
+    PYTHONUNBUFFERED=1 \
+    PIP_DISABLE_PIP_VERSION_CHECK=1 \
+    PATH="/venv/bin:$PATH"
+
+RUN apk upgrade --no-cache \
+	&& apk add --no-cache libjpeg-turbo \
     && pip install --upgrade pip \
-    && pip install --no-cache-dir -r requirements.txt \
-    && apk del .build-deps \
-    && apk add --no-cache libjpeg-turbo \
     && mkdir /data
+
+COPY --from=builder /venv /venv
 
 VOLUME ["/data"]
 
-ENV IG_USERNAME='' IG_PASSWORD=''
+ENV CACHE_FILE=/data/cache.json \
+    DATA_DIR=/data \
+    LOG_LEVEL=info \
+    ENABLE_ANALYTICS=true \
+    ENABLE_DEBUG=false \
+    ENABLE_UPLOAD=true \
+    IG_USER= \
+    IG_USERNAME= \
+    IG_PASSWORD=
 
-ENTRYPOINT ["python",  "/app/instagram_routine_bot.py", "-d", "-c", "/data/cache.json"]
+ADD ./app /app
 
-# docker run --name=instagram-bot -d -v ~/Documents/IG_DATA:/data -e IG_USERNAME='' -e IG_PASSWORD='' <img_id>
-
-
-# apk add build-base python-dev py-pip  zlib-dev
-# apk --no-cache --update-cache add 
-# pip install 
+ENTRYPOINT ["python",  "/app/instagram_routine_bot.py"]
