--- Dockerfile.multi
@@ -1,51 +1,0 @@
-FROM python:3.7-alpine3.10 AS builder
-
-WORKDIR /app
-
-ENV LANG=C.UTF-8
-ENV PYTHONUNBUFFERED=1
-# Turns off writing .pyc files
-ENV PYTHONDONTWRITEBYTECODE=1
-
-# Run from virtualenv
-ENV PATH="/venv/bin:$PATH"
-
-RUN apk upgrade --no-cache \
-	&& apk add --no-cache --virtual .build-deps build-base \
-	gcc gfortran libpng-dev openblas-dev jpeg-dev \
-    && pip install --upgrade pip
-
-RUN python -m venv /venv
-
-COPY *.py *.txt /app/
-
-RUN pip install --no-cache-dir -r requirements.txt
-
-
-
-FROM python:3.7-alpine3.10 
-
-WORKDIR /app
-
-# Extra python env
-ENV PYTHONDONTWRITEBYTECODE=1
-ENV PYTHONUNBUFFERED=1
-ENV PIP_DISABLE_PIP_VERSION_CHECK=1
-ENV PATH="/venv/bin:$PATH"
-
-RUN apk upgrade --no-cache \
-	&& apk add --no-cache libjpeg-turbo \
-    && pip install --upgrade pip \
-    && mkdir /data
-
-COPY --from=builder /venv /venv
-COPY *.py /app
-
-VOLUME ["/data"]
-
-ENV IG_USERNAME='' IG_PASSWORD=''
-
-ENTRYPOINT ["python",  "/app/instagram_routine_bot.py", "-d", "-c", "/data/cache.json"]
-
-
-# docker run --name=instagram-bot -d -v ~/Documents/IG_DATA:/data -e IG_USERNAME='' -e IG_PASSWORD='' <img_id>
