+++ doc/issues.md
@@ -1,0 +1,3 @@
+# Common Issues
+
+* If you encounter an issue `ClientError checkpoint_challenge_required`, that mean your access attempt has been flagged. Login manually to pass the required challenge.
