+++ requirements.txt
@@ -1,2 +1,3 @@
 instagram_private_api
 instagram_private_api_extensions
+ConfigArgParse
